requirements.txt
gunicorn==20.1.0
aioredis==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==3.1.0
cloudinary==1.33.0
constantly==15.1.0
cryptography==40.0.2
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.20
django-allauth==0.54.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-model-utils==4.3.1
django-smtp-ssl==1.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsonfield==3.1.0
msgpack==1.0.5
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.2.0
python-dotenv==0.21.1
python-magic==0.4.26
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
zope.interface==6.0
